fix all errors
