+++ src/tools/blast/components/results/BlastResultsButtons.tsx
@@ -17,12 +17,12 @@ import uniProtKBApiUrls, {
 
 import fetchData from '../../../../shared/utils/fetchData';
 
-import { PublicServerParameters } from '../../types/blastServerParameters';
+import { PublicServerParameters } from '../../../types/toolsServerParameters';
 import { Suggestions } from '../../../../uniprotkb/components/query-builder/AutocompleteWrapper';
 import { JobTypes } from '../../../types/toolsJobTypes';
 
 type ResubmitButtonProps = {
-  inputParamsData?: PublicServerParameters;
+  inputParamsData?: PublicServerParameters[JobTypes.BLAST];
 };
 
 const ResubmitButton: FC<ResubmitButtonProps> = ({ inputParamsData }) => {
@@ -38,32 +38,37 @@ const ResubmitButton: FC<ResubmitButtonProps> = ({ inputParamsData }) => {
     setDisabled(true);
 
     const taxonMapping = new Map();
-    const taxonRequests = (inputParamsData.taxids || '')
-      .split(',')
-      .map((id) => {
-        const idCleaned = id.trim();
-        if (!idCleaned) {
-          return;
-        }
-        // eslint-disable-next-line consistent-return
-        return fetchData<Suggestions>(
-          getSuggesterUrl(uniProtKBApiUrls.organismSuggester, idCleaned)
-        ).then((response) => {
-          const firstSuggestion = response?.data?.suggestions?.[0]?.value;
-          if (firstSuggestion) {
-            taxonMapping.set(
-              idCleaned,
-              `${firstSuggestion.replace(/ *\([^)]*\) */g, '')} [${idCleaned}]`
-            );
+    if ('taxids' in inputParamsData) {
+      const taxonRequests = (inputParamsData.taxids || '')
+        .split(',')
+        .map((id) => {
+          const idCleaned = id.trim();
+          if (!idCleaned) {
+            return;
           }
+          // eslint-disable-next-line consistent-return
+          return fetchData<Suggestions>(
+            getSuggesterUrl(uniProtKBApiUrls.organismSuggester, idCleaned)
+          ).then((response) => {
+            const firstSuggestion = response?.data?.suggestions?.[0]?.value;
+            if (firstSuggestion) {
+              taxonMapping.set(
+                idCleaned,
+                `${firstSuggestion.replace(
+                  / *\([^)]*\) */g,
+                  ''
+                )} [${idCleaned}]`
+              );
+            }
+          });
         });
-      });
 
-    try {
-      // wait up to 2 seconds to get the information
-      await Promise.race([Promise.all(taxonRequests), sleep(2000)]);
-    } catch (_) {
-      /* */
+      try {
+        // wait up to 2 seconds to get the information
+        await Promise.race([Promise.all(taxonRequests), sleep(2000)]);
+      } catch (_) {
+        /* */
+      }
     }
 
     const parameters = serverParametersToFormParameters(
@@ -91,7 +96,7 @@ const ResubmitButton: FC<ResubmitButtonProps> = ({ inputParamsData }) => {
 type BlastResultsButtonsProps = {
   jobId: string;
   selectedEntries: string[];
-  inputParamsData?: PublicServerParameters;
+  inputParamsData?: PublicServerParameters[JobTypes.BLAST];
 };
 
 const BlastResultsButtons: FC<BlastResultsButtonsProps> = ({
