+++ src/tools/blast/components/results/__tests__/BlastResultTable.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import BlastResultTable from '../BlastResultTable';
-import blastResultsMockData from '../../../__mocks__/server-jobs/example-truncated.json';
+import blastResultsMockData from '../../../../__mocks__/server-jobs/example-truncated.json';
 import renderWithRedux from '../../../../../shared/__test-helpers__/RenderWithRedux';
 import { fireEvent } from '@testing-library/react';
